NL-2-sql-App/pyproject.toml

[project]
name = "ai-sql-bot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "streamlit",
    "chromadb",
    "pydantic",
    "sqlparse",
    "python-dotenv",
    "pandas",
    "pytest",
    "reportlab>=4.4.3",
    "openai>=1.102.0",
    "google-generativeai>=0.8.5",
    "psutil>=7.0.0",
]
